refactor: extract GetTodos into custom hook and sub-components
Extract useTodos hook, TodoListItem, and TodoFilterDropdown from the
284-line GetTodos component, reducing it to ~40 lines.

Closes #9
